api/requirements.txt
flask==3.0.3
flask-cors==4.0.1
flask-migrate==4.0.7
flask-sqlalchemy==3.1.1
passlib==1.7.4
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
requests==2.32.2
sqlalchemy==2.0.30
email-validator==2.1.1
psycopg2-binary==2.9.9
py3dns==4.0.1